Split JWT verification functionality to the separate module making it available for using in plugins
