+++ src/postgraphile/withPostGraphileContext.ts
@@ -7,11 +7,7 @@ import { $$pgClient } from '../postgres/inventory/pgClientFromContext';
 import { pluginHookFromOptions } from './pluginHook';
 import { mixed, WithPostGraphileContextOptions } from '../interfaces';
 import { formatSQLForDebugging } from 'postgraphile-core';
-
-const undefinedIfEmpty = (
-  o?: Array<string | RegExp> | string | RegExp,
-): undefined | Array<string | RegExp> | string | RegExp =>
-  o && (!Array.isArray(o) || o.length) ? o : undefined;
+import jwtVerify, { jwtVerifyResult } from './jwtVerify';
 
 interface PostGraphileContext {
   [$$pgClient]: PoolClient;
@@ -309,89 +305,23 @@ async function getSettingsForPgClientTransaction({
   localSettings: Array<[string, string]>;
   jwtClaims: { [claimName: string]: mixed } | null;
 }> {
-  // Setup our default role. Once we decode our token, the role may change.
-  let role = pgDefaultRole;
-  let jwtClaims: { [claimName: string]: mixed } = {};
-
-  // If we were provided a JWT token, let us try to verify it. If verification
-  // fails we want to throw an error.
-  if (jwtToken) {
-    // Try to run `jwt.verify`. If it fails, capture the error and re-throw it
-    // as a 403 error because the token is not trustworthy.
-    try {
-      const jwtVerificationSecret = jwtPublicKey || jwtSecret;
-      // If a JWT token was defined, but a secret was not provided to the server or
-      // secret had unsupported type, throw a 403 error.
-      if (
-        !Buffer.isBuffer(jwtVerificationSecret) &&
-        typeof jwtVerificationSecret !== 'string' &&
-        typeof jwtVerificationSecret !== 'function'
-      ) {
-        // tslint:disable-next-line no-console
-        console.error(
-          `ERROR: '${
-            jwtPublicKey ? 'jwtPublicKey' : 'jwtSecret'
-          }' was not set to a string or buffer - rejecting JWT-authenticated request.`,
-        );
-        throw new Error('Not allowed to provide a JWT token.');
-      }
-
-      if (jwtAudiences != null && jwtVerifyOptions && 'audience' in jwtVerifyOptions)
-        throw new Error(
-          `Provide either 'jwtAudiences' or 'jwtVerifyOptions.audience' but not both`,
-        );
-
-      const claims = await new Promise((resolve, reject) => {
-        jwt.verify(
-          jwtToken,
-          jwtVerificationSecret,
-          {
-            ...jwtVerifyOptions,
-            audience:
-              jwtAudiences ||
-              (jwtVerifyOptions && 'audience' in (jwtVerifyOptions as object)
-                ? undefinedIfEmpty(jwtVerifyOptions.audience)
-                : ['postgraphile']),
-          },
-          (err, decoded) => {
-            if (err) reject(err);
-            else resolve(decoded);
-          },
-        );
-      });
-
-      if (typeof claims === 'string') {
-        throw new Error('Invalid JWT payload');
-      }
-
-      // jwt.verify returns `object | string`; but the `object` part is really a map
-      jwtClaims = claims as typeof jwtClaims;
-
-      const roleClaim = getPath(jwtClaims, jwtRole);
-
-      // If there is a `role` property in the claims, use that instead of our
-      // default role.
-      if (typeof roleClaim !== 'undefined') {
-        if (typeof roleClaim !== 'string')
-          throw new Error(
-            `JWT \`role\` claim must be a string. Instead found '${typeof jwtClaims['role']}'.`,
-          );
-
-        role = roleClaim;
-      }
-    } catch (error) {
-      // In case this error is thrown in an HTTP context, we want to add status code
-      // Note. jwt.verify will add a name key to its errors. (https://github.com/auth0/node-jsonwebtoken#errors--codes)
-      error.statusCode =
-        'name' in error && error.name === 'TokenExpiredError'
-          ? // The correct status code for an expired ( but otherwise acceptable token is 401 )
-            401
-          : // All other authentication errors should get a 403 status code.
-            403;
-
-      throw error;
-    }
-  }
+  const verifyResult = await new Promise((resolve, reject) => {
+    jwtVerify(
+      jwtToken,
+      jwtSecret,
+      jwtPublicKey,
+      jwtAudiences,
+      jwtRole,
+      jwtVerifyOptions,
+      pgDefaultRole,
+      (err, decoded) => {
+        if (err) reject(err);
+        else resolve(decoded);
+      },
+    );
+  });
+  let role = (verifyResult as jwtVerifyResult).role;
+  const jwtClaims = (verifyResult as jwtVerifyResult).jwtClaims;
 
   // Instantiate a map of local settings. This map will be transformed into a
   // Sql query.
@@ -560,23 +490,6 @@ export function debugPgClient(pgClient: PoolClient, allowExplain = false): PoolC
   }
 
   return pgClient;
-}
-
-/**
- * Safely gets the value at `path` (array of keys) of `inObject`.
- *
- * @private
- */
-function getPath(inObject: mixed, path: Array<string>): any {
-  let object = inObject;
-  // From https://github.com/lodash/lodash/blob/master/.internal/baseGet.js
-  let index = 0;
-  const length = path.length;
-
-  while (object && index < length) {
-    object = object[path[index++]];
-  }
-  return index && index === length ? object : undefined;
 }
 
 /**
